[AN] refactor(#237): 패키지 이름 수정
