+++ android/gitudy/presentation/src/main/java/com/takseha/presentation/firebase/MyFirebaseMessagingService.kt
@@ -11,7 +11,7 @@ import androidx.core.app.NotificationCompat
 import com.google.firebase.messaging.FirebaseMessaging
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
-import com.takseha.presentation.ui.login.SocialLoginCompleteActivity
+import com.takseha.presentation.ui.auth.SocialLoginCompleteActivity
 
 class MyFirebaseMessagingService : FirebaseMessagingService() {
     override fun onNewToken(token: String) {
