+++ android/gitudy/presentation/src/main/java/com/takseha/presentation/ui/auth/InputIdFragment.kt
@@ -1,4 +1,4 @@
-package com.takseha.presentation.ui.login
+package com.takseha.presentation.ui.auth
 
 import android.os.Bundle
 import android.text.Editable
