+++ android/gitudy/presentation/src/main/java/com/takseha/presentation/ui/auth/LoginActivity.kt
@@ -1,39 +1,37 @@
-package com.takseha.presentation.ui.login
+package com.takseha.presentation.ui.auth
 
 import android.content.Intent
-import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
 import androidx.core.content.ContextCompat
 import androidx.lifecycle.ViewModelProvider
 import com.takseha.presentation.R
-import com.takseha.presentation.databinding.ActivitySubLoginBinding
+import com.takseha.presentation.databinding.ActivityLoginBinding
 import com.takseha.presentation.viewmodel.LoginViewModel
 
-
-class SubLoginActivity : AppCompatActivity() {
-    private lateinit var binding: ActivitySubLoginBinding
+class LoginActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityLoginBinding
     private lateinit var viewModel: LoginViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_sub_login)
+        setContentView(R.layout.activity_login)
         window.statusBarColor = ContextCompat.getColor(this, R.color.BACKGROUND)
         setBinding()
 
         viewModel = ViewModelProvider(this)[LoginViewModel::class.java]
 
         with(binding) {
-            kakaoLoginBtn.setOnClickListener {
-                startLogin("KAKAO")
+            githubLoginBtn.setOnClickListener {
+                startLogin("GITHUB")
             }
-            googleLoginBtn.setOnClickListener {
-                startLogin("GOOGLE")
+            otherLoginBtn.setOnClickListener {
+                startActivity(Intent(baseContext, SubLoginActivity::class.java))
             }
         }
     }
-
     private fun setBinding() {
-        binding = ActivitySubLoginBinding.inflate(layoutInflater)
+        binding = ActivityLoginBinding.inflate(layoutInflater)
         val view = binding.root
         setContentView(view)
     }
