+++ android/gitudy/presentation/src/main/java/com/takseha/presentation/ui/auth/LoginWebViewActivity.kt
@@ -1,4 +1,4 @@
-package com.takseha.presentation.ui.login
+package com.takseha.presentation.ui.auth
 
 import android.content.Intent
 import android.net.Uri
