+++ android/gitudy/presentation/src/main/java/com/takseha/presentation/ui/auth/PopupDetail1Fragment.kt
@@ -1,4 +1,4 @@
-package com.takseha.presentation.ui.login
+package com.takseha.presentation.ui.auth
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
