+++ android/gitudy/presentation/src/main/res/layout/activity_login_webview.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.login.LoginWebViewActivity">
+    tools:context=".ui.auth.LoginWebViewActivity">
 
     <WebView
         android:id="@+id/loginWebView"
