+++ android/gitudy/presentation/src/main/res/layout/activity_popup_agreement.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.login.LoginActivity">
+    tools:context=".ui.auth.LoginActivity">
     <LinearLayout
         android:id="@+id/socialLoginCompleteTextBox"
         android:layout_width="wrap_content"
