+++ android/gitudy/presentation/src/main/res/layout/activity_social_login_complete.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/BACKGROUND"
-    tools:context=".ui.login.SocialLoginCompleteActivity">
+    tools:context=".ui.auth.SocialLoginCompleteActivity">
 
 
     <LinearLayout
