+++ android/gitudy/presentation/src/main/res/layout/fragment_input_id.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/BACKGROUND"
-    tools:context=".ui.login.InputIdFragment">
+    tools:context=".ui.auth.InputIdFragment">
 
     <TextView
         android:id="@+id/inputIdDescription1"
