+++ android/gitudy/presentation/src/main/res/layout/fragment_input_nickname.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/BACKGROUND"
-    tools:context=".ui.login.InputNicknameFragment">
+    tools:context=".ui.auth.InputNicknameFragment">
 
     <TextView
         android:id="@+id/inputNicknameDescription1"
