+++ android/gitudy/presentation/src/main/res/layout/fragment_popup_agreement.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.login.PopupAgreementFragment">
+    tools:context=".ui.auth.PopupAgreementFragment">
 
     <androidx.cardview.widget.CardView
         android:id="@+id/popupCardView"
