+++ android/gitudy/presentation/src/main/res/layout/fragment_popup_background.xml
@@ -4,5 +4,5 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/TRANSPARENT"
-    tools:context=".ui.login.PopupBackgroundFragment">
+    tools:context=".ui.auth.PopupBackgroundFragment">
 </FrameLayout>
