+++ android/gitudy/presentation/src/main/res/layout/fragment_popup_detail2.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/WHITE"
-    tools:context=".ui.login.PopupDetail2Fragment">
+    tools:context=".ui.auth.PopupDetail2Fragment">
 
     <LinearLayout
         android:layout_width="match_parent"
