+++ android/gitudy/presentation/src/main/res/layout/fragment_register_complete.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#15141b"
-    tools:context=".ui.login.RegisterCompleteFragment">
+    tools:context=".ui.auth.RegisterCompleteFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="wrap_content"
